Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 3653-xor-after-range-multiplication-queries-i/3653-xor-after-range-multiplication-queries-i.java
@@ -4,7 +4,7 @@ class Solution {
     }
 
     public int mySol_bf(int[] nums, int[][] queries) {
-        long mod = (long)1e9 + 7;
+        int mod = (int)1e9 + 7;
 
         for (int[] q : queries) {
             int index = q[0];
